pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max-line-length = 79

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
ignore_missing_imports = true
show_column_numbers = true
strict_optional = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude = "app/crud/association.py"
explicit_package_bases = true

[tool.poetry]
authors = ["Gustavo Mejia Quintero <admin@cocogum.org>"]
description = "Una aplicación para uso general en todas las comunidades."
name = "cocogum"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.13.2"
asyncio = "^3.4.3"
asyncpg = "^0.29.0"
fastapi = {extras = ["standard"], version = "^0.112.2"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
python = "^3.12.4"
python-multipart = "0.0.9"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pytest = "^8.2.2"
pytest-asyncio = "^0.20.3"
sqlalchemy = "^2.0.32"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
httpx = "^0.27.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.6.3"
taskipy = "^1.13.0"
pylint = "^3.2.7"
mypy = "^1.11.2"
autopep8 = "^2.3.1"

[tool.pylint]
ignore = ["venv"]
load-plugins = []
persistent = true

[tool.pylint.design]
max-args = 5
max-attributes = 7
max-branches = 12
max-locals = 15
max-parents = 7
max-returns = 6
max-statements = 50

[tool.pylint.format]
max-line-length = 79

[tool.pylint.imports]
known-first-party = []
known-standard-library = []
known-third-party = []

[tool.pylint.messages_control]
disable = ["C0114", "C0115", "C0116"]

[tool.pylint.reports]
output-format = "text"
reports = false

[tool.pylint.similarities]
min-similarity-lines = 4

[tool.pylint.typecheck]
ignored-classes = ["optparse.Values", "thread._local", "_thread._local", "argparse.Namespace"]

[tool.pylint.variables]
additional-builtins = []
allow-global-unused-variables = true

[tool.pyright]
exclude = ["./venv", "**/__pycache__"]
include = ["./app"]
pythonVersion = "3.12"

[tool.pytest.ini_options]
addopts = '-p no:warnings --cov=app --cov-report=term-missing --cov-report=html'
pythonpath = "."

[tool.ruff]
extend-exclude = ['alembic']
line-length = 79

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.taskipy.tasks]
format = 'ruff check . --fix && ruff format .'
lint = 'ruff check . && ruff check . --diff'
post_test = 'coverage html'
pre_test = 'task lint'
run = 'fastapi dev cocogum/app/main.py'
test = 'pytest --cov=cocogum -vv'